+++ eggs/Naaya/naaya/content/document/document_item.py
@@ -362,9 +362,9 @@ class NyDocument(document_item, NyAttributes, NyContainer, NyCheckControl,
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'commitVersion')
     def commitVersion(self, REQUEST=None):
         """ """
-        if ((not self.checkPermissionEditObject()) or
-            (self.checkout_user !=
-             self.REQUEST.AUTHENTICATED_USER.getUserName())):
+        user = self.REQUEST.AUTHENTICATED_USER.getUserName()
+        if (not self.checkPermissionEditObject()) or (
+            self.checkout_user != user):
             raise EXCEPTION_NOTAUTHORIZED, EXCEPTION_NOTAUTHORIZED_MSG
         if not self.hasVersion():
             raise EXCEPTION_NOVERSION, EXCEPTION_NOVERSION_MSG
@@ -374,6 +374,7 @@ class NyDocument(document_item, NyAttributes, NyContainer, NyCheckControl,
         self.version = None
         self._p_changed = 1
         self.recatalogNyObject(self)
+        notify(NyContentObjectEditEvent(self, user))
         if REQUEST: REQUEST.RESPONSE.redirect('%s/index_html' %
                                                 self.absolute_url())