+++ eggs/Naaya/naaya/content/file/file_item.py
@@ -403,7 +403,9 @@ class NyFile_extfile(file_item, NyAttributes, NyItem, NyFolderishVersioning, NyC
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'commitVersion')
     def commitVersion(self, REQUEST=None):
         """ """
-        if (not self.checkPermissionEditObject()) or (self.checkout_user != self.REQUEST.AUTHENTICATED_USER.getUserName()):
+        user = self.REQUEST.AUTHENTICATED_USER.getUserName()
+        if (not self.checkPermissionEditObject()) or (
+            self.checkout_user != user):
             raise EXCEPTION_NOTAUTHORIZED, EXCEPTION_NOTAUTHORIZED_MSG
         if not self.hasVersion():
             raise EXCEPTION_NOVERSION, EXCEPTION_NOVERSION_MSG
@@ -414,6 +416,7 @@ class NyFile_extfile(file_item, NyAttributes, NyItem, NyFolderishVersioning, NyC
         self.version = None
         self._p_changed = 1
         self.recatalogNyObject(self)
+        notify(NyContentObjectEditEvent(self, user))
         if REQUEST: REQUEST.RESPONSE.redirect('%s/index_html' % self.absolute_url())
 
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'startVersion')