+++ eggs/Naaya/naaya/content/geopoint/geopoint_item.py
@@ -278,7 +278,9 @@ class NyGeoPoint(geopoint_item, NyAttributes, NyItem, NyCheckControl, NyContentT
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'commitVersion')
     def commitVersion(self, REQUEST=None):
         """ """
-        if (not self.checkPermissionEditObject()) or (self.checkout_user != self.REQUEST.AUTHENTICATED_USER.getUserName()):
+        user = self.REQUEST.AUTHENTICATED_USER.getUserName()
+        if (not self.checkPermissionEditObject()) or (
+            self.checkout_user != user):
             raise EXCEPTION_NOTAUTHORIZED, EXCEPTION_NOTAUTHORIZED_MSG
         if not self.hasVersion():
             raise EXCEPTION_NOVERSION, EXCEPTION_NOVERSION_MSG
@@ -288,6 +290,7 @@ class NyGeoPoint(geopoint_item, NyAttributes, NyItem, NyCheckControl, NyContentT
         self.version = None
         self._p_changed = 1
         self.recatalogNyObject(self)
+        notify(NyContentObjectEditEvent(self, user))
         if REQUEST: REQUEST.RESPONSE.redirect('%s/index_html' % self.absolute_url())
 
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'startVersion')