+++ eggs/Naaya/naaya/content/story/story_item.py
@@ -382,7 +382,9 @@ class NyStory(story_item, NyAttributes, NyContainer, NyCheckControl, NyContentTy
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'commitVersion')
     def commitVersion(self, REQUEST=None):
         """ """
-        if (not self.checkPermissionEditObject()) or (self.checkout_user != self.REQUEST.AUTHENTICATED_USER.getUserName()):
+        user = self.REQUEST.AUTHENTICATED_USER.getUserName()
+        if (not self.checkPermissionEditObject()) or (
+            self.checkout_user != user):
             raise EXCEPTION_NOTAUTHORIZED, EXCEPTION_NOTAUTHORIZED_MSG
         if not self.hasVersion():
             raise EXCEPTION_NOVERSION, EXCEPTION_NOVERSION_MSG
@@ -393,6 +395,7 @@ class NyStory(story_item, NyAttributes, NyContainer, NyCheckControl, NyContentTy
         self.version = None
         self._p_changed = 1
         self.recatalogNyObject(self)
+        notify(NyContentObjectEditEvent(self, user))
         if REQUEST: REQUEST.RESPONSE.redirect('%s/index_html' % self.absolute_url())
 
     security.declareProtected(PERMISSION_EDIT_OBJECTS, 'startVersion')